### 2023-09-08 #5816 @editorialbot
[PRE REVIEW]: sparse-lm: Sparse linear regression models in Python
**Submitting author:** @lbluque (<a href="http://orcid.org/0000-0002-6453-9545">Luis Barroso-Luque</a>)
**Repository:** https://github.com/CederGroupHub/sparse-lm
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.5.1
