### 2023-09-11 #5816 @jbytecode
@arfon - I think some chapters of the manuscript is empty and the paper is not linked to the citation entries defined in the bibtex file so this submission needs some attention before the review process, or a query-scope label can be assigned.
