### 2023-09-12 #5816 @lbluque
@jbytecode thanks a lot for pointing this out! I have updated the paper branch and the issue should be fixed now.
