+++ .github/release.yml
@@ -1,0 +1,22 @@
+changelog:
+  exclude:
+    authors: [dependabot, github-actions, pre-commit-ci]
+  categories:
+    - title: 🎉 New Features
+      labels: [feature]
+    - title: 🐛 Bug Fixes
+      labels: [fix]
+    - title: 🛠 Enhancements
+      labels: [enhancement]
+    - title: 📖 Documentation
+      labels: [docs]
+    - title: 💡 Refactoring
+      labels: [refactor]
+    - title: 🧪 Tests
+      labels: [tests]
+    - title: 💥 Breaking Changes
+      labels: [breaking]
+    - title: 🔒 Security Fixes
+      labels: [security]
+    - title: 🤷‍♂️ Other Changes
+      labels: ["*"]
