### 2023-09-13 b1aea99
DEV: remove python 3.8 tests
