+++ .github/workflows/test.yml
@@ -18,7 +18,7 @@ jobs:
     strategy:
       max-parallel: 10
       matrix:
-        python_version: [3.9", "3.10", "3.11"]
+        python_version: ["3.9", "3.10", "3.11"]
 
     steps:
       - uses: actions/checkout@v4
