### 2023-09-13 a2da37b
remove printing sklearn info
