### 2023-09-13 f7f4745
DEV: add release workflow
