+++ .github/workflows/test.yml
@@ -12,6 +12,8 @@ on:
     branches:
       - main
 
+  workflow_call:
+
 jobs:
   test:
     runs-on: ubuntu-20.04
