+++ src/sparselm/model/_miqp/_best_subset.py
@@ -11,8 +11,9 @@ from numbers import Real
 from types import SimpleNamespace
 from typing import Any
 
+import numpy as np
 import cvxpy as cp
-from numpy.typing import ArrayLike, NDArray
+from numpy.typing import NDArray, NDArray
 from sklearn.utils._param_validation import Interval
 
 from sparselm.model._base import TikhonovMixin
@@ -26,7 +27,7 @@ class BestSubsetSelection(MIQPl0):
     Generalized best subset that allows grouping subsets.
 
     Args:
-        groups (ArrayLike):
+        groups (NDArray):
             array-like of integers specifying groups. Length should be the
             same as model, where each integer entry specifies the group
             each parameter corresponds to. If no grouping is required,
@@ -98,7 +99,7 @@ class BestSubsetSelection(MIQPl0):
 
     def __init__(
         self,
-        groups: ArrayLike | None = None,
+        groups: NDArray | None = None,
         sparse_bound=100,
         big_M: int = 100,
         hierarchy: list[list[int]] | None = None,
@@ -124,12 +125,12 @@ class BestSubsetSelection(MIQPl0):
 
     def _generate_constraints(
         self,
-        X: ArrayLike,
-        y: ArrayLike,
+        X: NDArray,
+        y: NDArray,
         beta: cp.Variable,
         parameters: SimpleNamespace | None = None,
         auxiliaries: SimpleNamespace | None = None,
-    ) -> list[cp.constraints]:
+    ) -> list[cp.Constraint]:
         """Generate the constraints for best subset selection."""
         constraints = super()._generate_constraints(X, y, beta, parameters, auxiliaries)
         constraints += [cp.sum(auxiliaries.z0) <= parameters.sparse_bound]
@@ -140,7 +141,7 @@ class RidgedBestSubsetSelection(TikhonovMixin, BestSubsetSelection):
     r"""MIQP best subset selection Regressor with Ridge/Tihkonov regularization.
 
     Args:
-        groups (ArrayLike):
+        groups (NDArray):
             array-like of integers specifying groups. Length should be the
             same as model, where each integer entry specifies the group
             each parameter corresponds to. If no grouping is required,
@@ -216,12 +217,12 @@ class RidgedBestSubsetSelection(TikhonovMixin, BestSubsetSelection):
 
     def __init__(
         self,
-        groups: ArrayLike | None = None,
+        groups: NDArray | None = None,
         sparse_bound: int = 100,
         eta: float = 1.0,
         big_M: int = 100,
         hierarchy: list[list[int]] | None = None,
-        tikhonov_w: NDArray[float] | None = None,
+        tikhonov_w: NDArray[np.floating] | None = None,
         ignore_psd_check: bool = True,
         fit_intercept: bool = False,
         copy_X: bool = True,
