+++ src/sparselm/model/_miqp/_regularized_l0.py
@@ -28,7 +28,7 @@ from typing import Any
 
 import cvxpy as cp
 import numpy as np
-from numpy.typing import ArrayLike, NDArray
+from numpy.typing import NDArray, NDArray
 from sklearn.utils._param_validation import Interval
 
 from sparselm.model._base import TikhonovMixin
@@ -52,7 +52,7 @@ class RegularizedL0(MIQPl0):
     valued slack variables.
 
     Args:
-        groups (ArrayLike):
+        groups (NDArray):
             1D array-like of integers specifying groups. Length should be the
             same as model, where each integer entry specifies the group
             each parameter corresponds to. If no grouping is needed pass a list
@@ -119,7 +119,7 @@ class RegularizedL0(MIQPl0):
 
     def __init__(
         self,
-        groups: ArrayLike | None = None,
+        groups: NDArray | None = None,
         alpha: float = 1.0,
         big_M: int = 100,
         hierarchy: list[list[int]] | None = None,
@@ -145,8 +145,8 @@ class RegularizedL0(MIQPl0):
 
     def _generate_objective(
         self,
-        X: ArrayLike,
-        y: ArrayLike,
+        X: NDArray,
+        y: NDArray,
         beta: cp.Variable,
         parameters: SimpleNamespace | None = None,
         auxiliaries: SimpleNamespace | None = None,
@@ -169,7 +169,7 @@ class MixedL0(RegularizedL0, metaclass=ABCMeta):
 
     def __init__(
         self,
-        groups: ArrayLike | None = None,
+        groups: NDArray | None = None,
         alpha: float = 1.0,
         eta: float = 1.0,
         big_M: int = 100,
@@ -184,7 +184,7 @@ class MixedL0(RegularizedL0, metaclass=ABCMeta):
         """Initialize Regressor.
 
         Args:
-            groups (ArrayLike):
+            groups (NDArray):
                 1D array-like of integers specifying groups. Length should be the
                 same as model, where each integer entry specifies the group
                 each parameter corresponds to. If no grouping is needed pass a list
@@ -242,8 +242,8 @@ class MixedL0(RegularizedL0, metaclass=ABCMeta):
     @abstractmethod
     def _generate_objective(
         self,
-        X: ArrayLike,
-        y: ArrayLike,
+        X: NDArray,
+        y: NDArray,
         beta: cp.Variable,
         parameters: SimpleNamespace | None = None,
         auxiliaries: SimpleNamespace | None = None,
@@ -276,7 +276,7 @@ class L1L0(MixedL0):
     valued slack variables.
 
     Args:
-        groups (ArrayLike):
+        groups (NDArray):
             1D array-like of integers specifying groups. Length should be the
             same as model, where each integer entry specifies the group
             each parameter corresponds to. If no grouping is needed pass a list
@@ -340,7 +340,7 @@ class L1L0(MixedL0):
 
     def __init__(
         self,
-        groups: ArrayLike | None = None,
+        groups: NDArray | None = None,
         alpha: float = 1.0,
         eta: float = 1.0,
         big_M: int = 100,
@@ -367,7 +367,7 @@ class L1L0(MixedL0):
         )
 
     def _generate_auxiliaries(
-        self, X: ArrayLike, y: ArrayLike, beta: cp.Variable, parameters: SimpleNamespace
+        self, X: NDArray, y: NDArray, beta: cp.Variable, parameters: SimpleNamespace
     ) -> SimpleNamespace | None:
         """Generate the boolean slack variable."""
         auxiliaries = super()._generate_auxiliaries(X, y, beta, parameters)
@@ -377,12 +377,12 @@ class L1L0(MixedL0):
 
     def _generate_constraints(
         self,
[…]
