+++ .pre-commit-config.yaml
@@ -18,7 +18,7 @@ repos:
     args: ['--maxkb=500']
 
 - repo: https://github.com/psf/black
-  rev: 23.7.0
+  rev: 23.9.1
   hooks:
   - id: black
 
