+++ .github/workflows/build.yml
@@ -4,7 +4,7 @@ on: [workflow_dispatch, workflow_call]
 
 jobs:
 
-  build_sdist:
+  build-sdist:
     name: Build sdist
     runs-on: ubuntu-latest
 
@@ -13,7 +13,7 @@ jobs:
         with:
           fetch-depth: 0 # Optional, use if you use setuptools_scm
 
-      - name: Build sdist
+      - name: Build
         run: pipx run build --sdist
 
       - uses: actions/upload-artifact@v3
