### 2023-09-15 cf3730c
add code_of_conduct and update issue templates
