+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -43,7 +43,7 @@ body:
     id: terms
     attributes:
       label: Code of Conduct
-      description: By submitting this issue, you agree to follow our [Code of Conduct](https://github.com/materialsproject/.github/blob/main/.github/code_of_conduct.md)
+      description: By submitting this issue, you agree to follow our [Code of Conduct](https://github.com/CederGroupHub/sparse-lm/blob/main/code_of_conduct.md)
       options:
         - label: I agree to follow this project's Code of Conduct
           required: true
