### 2023-09-16 #5816 @jbytecode
@lbluque - Thank you for sending your work to JOSS. I am the handling editor for this submission.

Do you have suggestions for the potential reviewers? You can use the list https://reviewers.joss.theoj.org/lookup to suggest reviewers. If so, please mention their GitHub usernames here without using the `@` character as we don't to bother them with automatic notifications.

Thank you in advance.
