### 2023-09-18 #5816 @jbytecode
👋👋👋 Dear @ejolly, @potterzot, @mmrabe 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**sparse-lm: Sparse linear regression models in Python**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5816).

Thank you in advance!
