### 2023-09-18 f023eb4
STY: fix more mypy errors
