### 2023-09-20 e4ff4c4
DEV: require build only to release
