+++ .github/workflows/release.yml
@@ -29,7 +29,7 @@ jobs:
     secrets: inherit
 
   release-pypi:
-    needs: [build, docs]
+    needs: [build]
     runs-on: ubuntu-latest
     if: github.event.inputs.release-pypi == 'true'
 
