### 2023-09-21 c2c4fd3
STY: more typing fixes
