### 2023-09-21 #5816 @mhu48
Hi,

Thank you for the invitation. I would love to serve as a reviewer for the
paper as well!

Best,

Ming

On Wed, Sep 20, 2023 at 1:12 PM Mehmet Hakan Satman <
***@***.***> wrote:

> 👋👋👋 Dear @mhu48 <https://github.com/mhu48>, @htjb
> <https://github.com/htjb> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled: "*sparse-lm: Sparse linear
> regression models in Python*". You can find more information at the top
> of this Github issue (#5816
> <https://github.com/openjournals/joss-reviews/issues/5816>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
