### 2023-09-27 9dad2bd
Delete .github/ISSUE_TEMPLATE directory
