+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -1,49 +1,0 @@
-name: Feature Request
-description: Propose a new idea for this project
-title: "[Feature Request]: "
-labels: ["feature_request", "triage"]
-body:
-  - type: markdown
-    attributes:
-      value: |
-        Thanks for taking the time to submit this feature request!
-  - type: input
-    id: contact
-    attributes:
-      label: Email (Optional)
-      description: How can we get in touch with you if we need more info?
-      placeholder: ex. email@example.com
-    validations:
-      required: false
-  - type: textarea
-    id: problem
-    attributes:
-      label: Problem
-      description: Briefly describe the problem you're experiencing.
-      placeholder: E.g., I would like to be able to do X.
-    validations:
-      required: true
-  - type: textarea
-    id: solution
-    attributes:
-      label: Proposed Solution
-      description: Detail the feature or solution you'd like to see implemented.
-      placeholder: E.g., Add a module Y with these functionalities (1)..., (2)...
-    validations:
-      required: true
-  - type: textarea
-    id: alternatives
-    attributes:
-      label: Alternatives
-      description: List any alternative solutions, prior art or features you have considered, if any.
-      placeholder: E.g., Similar functionality is found in package <Y>.
-    validations:
-      required: false
-  - type: checkboxes
-    id: terms
-    attributes:
-      label: Code of Conduct
-      description: By submitting this issue, you agree to follow our [Code of Conduct](https://github.com/CederGroupHub/sparse-lm/blob/main/code_of_conduct.md)
-      options:
-        - label: I agree to follow this project's Code of Conduct
-          required: true
