### 2023-09-27 b7169d3
DEV: inherit secrets for test
