+++ .github/workflows/release.yml
@@ -18,6 +18,7 @@ on:
 jobs:
   test:
     uses: ./.github/workflows/test.yml
+    secrets: inherit
 
   build:
     needs: test
