### 2023-09-27 7a66993
lint and add mypy to pre-commit
