+++ pyproject.toml
@@ -81,9 +81,6 @@ omit = ["*/__init__.py"]
 convention = "google"
 add_ignore = ["D107"]
 
-[tool.mypy]
-plugins = ["numpy.typing.mypy_plugin"]
-
 [[tool.mypy.overrides]]
 module = ["sklearn.*", "scipy.linalg"]
 ignore_missing_imports = true
