### 2023-09-27 4e12fb2
fix NDArray types
