+++ src/sparselm/_utils/validation.py
@@ -1,5 +1,7 @@
 """Data and hyper-parameters validation utilities."""
 
+from __future__ import annotations
+
 import numpy as np
 from numpy.typing import NDArray
 
