+++ .pre-commit-config.yaml
@@ -38,7 +38,7 @@ repos:
     - --profile=black
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v3.10.1
+  rev: v3.13.0
   hooks:
     - id: pyupgrade
       args: [--py38-plus]
@@ -71,6 +71,6 @@ repos:
   - id: rst-inline-touching-normal
 
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: ''  # Use the sha / tag you want to point at
+    rev: 'v1.5.1'  # Use the sha / tag you want to point at
     hooks:
     -   id: mypy
