+++ .github/workflows/test.yml
@@ -45,9 +45,9 @@ jobs:
         run: |
           pytest tests --cov=sparselm --cov-report=xml
 
-      - if: matrix.python_version == '3.11'
-        name: upload coverage report
-        uses: actions/upload-artifact@v3
+      - if: ${{ matrix.python_version == 3.11 && github.event_name == 'push' }}
+        name: codacy-coverage-reporter
+        uses: codacy/codacy-coverage-reporter-action@v1
         with:
-          name: coverage
-          path: coverage.xml
+            project-token: ${{ secrets.CODACY_PROJECT_TOKEN }}
+            coverage-reports: coverage.xml
