### 2023-10-02 #5867 @htjb
@lbluque this is a nice piece of code that I believe fills a gap in the capabilities of sklearn. The documentation is well written and formatted as is the paper.

I had a few comments on the code and paper below.

## Paper

- I think in line 56 there is an unrendered symbol between `and` and `are`.
- Are you aware of any particular examples of `sparse-lm` being used in the literature? If so it might be good to reference these somewhere in the paper.
- You mention a few different packages that can be used to fit some specific sparse linear regression models. My understanding is that `sparse-lm` aims to be a catch all for the things missing in `scikit-learn` so that the user only needs to install one package in addition to `sklearn`. In the paper where you list the implemented regression models and model selection tools I think it would be good to add which packages currently feature these models e.g. (`celer` and `grupyr`) next to Lasso and Group Lasso. This will help to emphasis the benefits of having all of the models in one place and identify any gaps in the existing toolbox.
- I think it would be nice to see a physically motivated example/use case in the paper just to emphasis how important the tool is.
- I think `nore` on line 45 should be `nor`.

## Code

- Could you please clarify for my benefit why you have reimplemented the `GridSearchCV` functionality in `sklearn`? You use the `sklearn` grid search in the readme example.
- It appears that the example in the readme crashes. This should ideally be fixed.
- There is a nice guide for contributing to the code in the documentation but it would be good to have this information in the README to since this is the first port of call for most people before looking at the documentation. I would just add a section saying contributions are welcome and link the documentation page to the readme.
- Similarly I would add install instructions to the README.
- I would add a list of requirements to the readme and maybe a `requirements.txt`. I'm aware that they are listed in the `pyproject.toml` but this is harder to parse than a `requirements.txt`.
- The API documentation is very thorough but there are a few formatting issues that it would be good to check over. e.g for `sparselm.model.RidgedBestSubsetSelection` under `VARIABLES:` and a few of the other models. Similarly in `GridSearchCV`.
