### 2023-10-02 ec70b6e
fix/improve README example
