+++ README.md
@@ -34,9 +34,9 @@ from sklearn.datasets import make_regression
 from sklearn.model_selection import GridSearchCV
 from sparselm.model import AdaptiveLasso
 
-X, y = make_regression(n_samples=200, n_features=5000, random_state=0)
+X, y = make_regression(n_samples=100, n_features=80, n_informative=10, random_state=0)
 alasso = AdaptiveLasso(fit_intercept=False)
-param_grid = {'alpha': np.logsppace(-7, -2)}
+param_grid = {'alpha': np.logspace(-8, 2, 10)}
 
 cvsearch = GridSearchCV(alasso, param_grid)
 cvsearch.fit(X, y)
