### 2023-10-02 5c10e97
Add install and contributing sections in README
