+++ README.md
@@ -23,6 +23,16 @@ Available regression models
 - Adaptive versions of Lasso, Group Lasso, Overlap Group Lasso, Sparse Group Lasso & Ridged Group Lasso.
 - Best Subset Selection, Ridged Best Subset, L0, L1L0 & L2L0 (all with optional grouping of parameters)
 
+Installation
+------------
+**sparse-lm** is available on [PyPI](https://pypi.org/project/sparse-lm/), and can be installed via pip:
+
+```bash
+pip install sparse-lm
+```
+
+Additional information on installation can be found in the [here](https://cedergrouphub.github.io/sparse-lm/install.html).
+
 Basic usage
 -----------
 If you already use **scikit-learn**, using **sparse-lm** will be very easy. Just use any
@@ -43,4 +53,12 @@ cvsearch.fit(X, y)
 print(cvsearch.best_params_)
 ```
 
-For more details on use and functionality see the [documentation](https://cedergrouphub.github.io/sparse-lm/).
+For more details on use and functionality have a look at the
+[examples](https://cedergrouphub.github.io/sparse-lm/auto_examples/index.html) and
+[API](https://cedergrouphub.github.io/sparse-lm/api.html) sections of the documentation.
+
+Contributing
+------------
+
+We welcome any contributions that you think may improve the package! Please have a look at the
+[contribution guidelines](https://cedergrouphub.github.io/sparse-lm/contributing.html) in the documentation.
