### 2023-10-03 cc75a9f
add table to paper
