### 2023-10-08 #5867 @jbytecode
@lbluque - Could you please set the license of the software explicitly as defined in https://joss.readthedocs.io/en/latest/submitting.html
