### 2023-10-08 #5867 @lbluque
> @lbluque - Could you please set the license of the software explicitly as defined in https://joss.readthedocs.io/en/latest/submitting.html. The license should be a in [OSI-approved license](https://opensource.org/licenses)

Hi @jbytecode. I am not sure what you mean by setting it explicitly. We have a BSD-3-clause-lbnl license, which is [osi approved](https://opensource.org/license/bsd-3-clause-lbnl/) in our LICENSE file.

Can you please point out if I have missed anything?
