### 2023-10-09 #5867 @lbluque
Ah, got it! Unfortunately, I think Github can't parse our specific license to have the link mention 3-BSD explicitly.
