+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ ci:
 
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v4.4.0
+  rev: v4.5.0
   hooks:
   - id: check-yaml
   - id: fix-encoding-pragma
@@ -38,7 +38,7 @@ repos:
     - --profile=black
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v3.13.0
+  rev: v3.15.0
   hooks:
     - id: pyupgrade
       args: [--py38-plus]
@@ -71,6 +71,6 @@ repos:
   - id: rst-inline-touching-normal
 
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v1.5.1'  # Use the sha / tag you want to point at
+    rev: 'v1.6.0'  # Use the sha / tag you want to point at
     hooks:
     -   id: mypy
