+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy >=1.23
+cvxpy >=1.2
+scikit-learn > 1.2
+scipy >=1.9
+joblib
