### 2023-10-24 #5867 @lbluque
I finally had some time to make edits. Thanks for your patience and your thorough review @htjb!

> ## Paper
>
>     * I think in line 56 there is an unrendered symbol between `and` and `are`.
Fixed, thanks.

>     * Are you aware of any particular examples of `sparse-lm` being used in the literature? If so it might be good to reference these somewhere in the paper.
I am only aware of our own work. I have added some citations at the end of the end of the Statement of Need section (though it feels a bit like a self-plug :electric_plug: )

>     * You mention a few different packages that can be used to fit some specific sparse linear regression models. My understanding is that `sparse-lm` aims to be a catch all for the things missing in `scikit-learn` so that the user only needs to install one package in addition to `sklearn`. In the paper where you list the implemented regression models and model selection tools I think it would be good to add which packages currently feature these models e.g. (`celer` and `grupyr`) next to Lasso and Group Lasso. This will help to emphasis the benefits of having all of the models in one place and identify any gaps in the existing toolbox.
Great suggestion! I added a table in the usage section that makes it much more clear what is and isn't available in the different packages.

>     * I think it would be nice to see a physically motivated example/use case in the paper just to emphasis how important the tool is.
I personally prefer to exclude direct examples in JOSS papers to keep the manuscript short and concise. And especially for a package like `sparse-lm` with applicability in a very wide range of fields, I think it can also over-emphasize a specific field. In lieu, I added citations to many applications in various fields of regression with structured sparsity in the Summary and Statement of need sections. Please let me know your thoughts.

>     * I think `nore` on line 45 should be `nor`.
Fixed thank you!

>
> ## Code
>
>     * Could you please clarify for my benefit why you have reimplemented the `GridSearchCV` functionality in `sklearn`? You use the `sklearn` grid search in the readme example.
The only reason is that the `GridSearchCV` implemented in `sparse-lm` allows to automatically pick hyperparameters based on the "1-std rule" (1 std from the optimum, and those favor sparsity over CV error), rather thank the optimal value given by the `sklearn` version, otherwise they are exactly the same.

>     * It appears that the example in the readme crashes. This should ideally be fixed.
Fixed as well. Thanks a lot!

>     * There is a nice guide for contributing to the code in the documentation but it would be good to have this information in the README to since this is the first port of call for most people before looking at the documentation. I would just add a section saying contributions are welcome and link the documentation page to the readme.
Very good point. Added links to README to make it easier for users and devs.

>     * Similarly I would add install instructions to the README.
Same as above.

>     * I would add a list of requirements to the readme and maybe a `requirements.txt`. I'm aware that they are listed in the `pyproject.toml` but this is harder to parse than a `requirements.txt`.
Added a requirements.txt to the repo as well.

>     * The API documentation is very thorough but there are a few formatting issues that it would be good to check over. e.g for `sparselm.model.RidgedBestSubsetSelection` under `VARIABLES:` and a few of the other models. Similarly in `GridSearchCV`.
Thanks a lot for cathing these. I went in and think I have now fixed all these formatting issues.

Please have a look at these changes when you get a chance, and let us know if you have further comments or suggestions. Thanks again!
