+++ paper/paper.md
@@ -163,18 +163,18 @@ options are implemented. The implemented models are listed below:
 The table below shows the regression models that are implemented in `sparse-lm` as well
 as available implementations in other Python packages. $\checkmark$ indicates that the
 
-|             Model             |  `sparse-lm` |    `celer` |   `groupyr` | `group-lasso` |    `skglm`   |  `abess` |
-|:-----------------------------:|:------------:|:---------:|:-----------:|:-----------:|:------------:|:--------:|
-|       (Adaptive) Lasso        | $\checkmark$️ | $\checkmark$️ |             |             | $\checkmark$️ |    ️     |
-|    (Adaptive) Group Lasso     | $\checkmark$️ | $\checkmark$️ | $\checkmark$️ | $\checkmark$️ | $\checkmark$ |    ️     |
-| (Adaptive) Sparse Group Lasso | $\checkmark$️ |           | $\checkmark$️ | $\checkmark$️ | $\checkmark$ |    ️     |
-| (Adaptive) Ridged Group Lasso | $\checkmark$️ |           |             |             | $\checkmark$ |          |
-|     Best Subset Selection     | $\checkmark$️ |           |             |             |              |    ️     |
-| Ridged Best Subset Selection  | $\checkmark$️ |           |             |             |              |    ️     |
-|     $\ell_0$ pseudo-norm      | $\checkmark$️ |           |             |             |              |    ️     |
-|   $\ell_0\ell_2$ mixed-norm   | $\checkmark$️ |           |             |             |              |          |
-|   $\ell_{1/2}$ psuedo-norm    |              |           |             |             | $\checkmark$ |          |
-|   $\ell_{2/3}$ psuedo-norm    |              |           |             |             | $\checkmark$ |          |
+|             Model             |  `sparse-lm` |    `celer` |   `groupyr` | `group-lasso` |    `skglm`   |    `abess`   |
+|:-----------------------------:|:------------:|:---------:|:-----------:|:-----------:|:------------:|:------------:|
+|       (Adaptive) Lasso        | $\checkmark$️ | $\checkmark$️ |             |             | $\checkmark$️ |       ️      |
+|    (Adaptive) Group Lasso     | $\checkmark$️ | $\checkmark$️ | $\checkmark$️ | $\checkmark$️ | $\checkmark$ |       ️      |
+| (Adaptive) Sparse Group Lasso | $\checkmark$️ |           | $\checkmark$️ | $\checkmark$️ | $\checkmark$ |       ️      |
+| (Adaptive) Ridged Group Lasso | $\checkmark$️ |           |             |             | $\checkmark$ |              |
+|     Best Subset Selection     | $\checkmark$️ |           |             |             |              | ️$\checkmark$ |
+| Ridged Best Subset Selection  | $\checkmark$️ |           |             |             |              |       ️      |
+|     $\ell_0$ pseudo-norm      | $\checkmark$️ |           |             |             |              |       ️      |
+|   $\ell_0\ell_2$ mixed-norm   | $\checkmark$️ |           |             |             |              |              |
+|   $\ell_{1/2}$ psuedo-norm    |              |           |             |             | $\checkmark$ |              |
+|   $\ell_{2/3}$ psuedo-norm    |              |           |             |             | $\checkmark$ |              |
 
 Note that only `sparse-lm` includes adaptive versions of Lasso based estimators. However, some of the third party packages,
 notably `skglm` and `abess`, include additional penalties and regression objectives that are not implemented in `sparse-lm`.
