### 2023-10-31 #5867 @htjb
Hi @lbluque, the changes look good! The table is a nice addition to the paper and the readme is looking good.

I can understand your point about adding examples in to the paper. I think referencing different fields where sparse linear regression models are being used is a reasonable alternative and referencing a paper where the package has already been used in the statement of need is great.

The package is a nice addition to the python toolbox! @jbytecode I ticked off my checklist. Please let me know if there is anything else I need to do.

@lbluque neat work!
