+++ .pre-commit-config.yaml
@@ -18,7 +18,7 @@ repos:
     args: ['--maxkb=500']
 
 - repo: https://github.com/psf/black
-  rev: 23.9.1
+  rev: 23.10.1
   hooks:
   - id: black
 
@@ -71,6 +71,6 @@ repos:
   - id: rst-inline-touching-normal
 
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v1.6.0'  # Use the sha / tag you want to point at
+    rev: 'v1.6.1'  # Use the sha / tag you want to point at
     hooks:
     -   id: mypy
