### 2023-11-19 #5867 @mhu48
Thank you fo the reminder, apparently something interesting happened last
week and I thought I already submitted the changes on github. Nevertheless,
I have updated the list just now. Please let me know if it is still not
reflected.

On a side note, how do I know how many papers I have reviewed for JOSS so
far? I tried searching for email chains but that was quite a mess. Thanks
so much for your help in advance.

Best,

Mign

On Sun, Nov 19, 2023 at 11:47 AM Mehmet Hakan Satman <
***@***.***> wrote:

> @mhu48 <https://github.com/mhu48> - Could you please update your status
> and inform us on how is your review going? Thank you in advance.
>
