### 2023-11-20 #5867 @mhu48
Hi,

Thank you for the reminder. I just saw your reply. Please find my report
below.

"I'm pleased to report that the author has diligently addressed the
concerns raised during the previous review process, and the changes made
have significantly improved the overall quality of the package.

Specifically, the updated README file is now well-structured, providing
clear and comprehensive documentation. The author has effectively
incorporated the necessary information, making it accessible and
user-friendly.

The concerns I previously raised and in the list of reviewer questions have
been thoughtfully addressed, and the modifications implemented are not only
acceptable but enhance the clarity and usability of the package. The
inclusion of additional details has notably improved the overall robustness
of the code.

In conclusion, this is a nice addition to the Python toolbox, offering
valuable functionality for sparse linear regression models."

Best,

Ming

On Sun, Nov 19, 2023 at 12:02 PM Mehmet Hakan Satman <
***@***.***> wrote:

> @mhu48 <https://github.com/mhu48> - I can see you've check all of the
> tasks in your check list, however, it would be nice to have a review report
> of you here.
>
