+++ paper/paper.bib
@@ -192,7 +192,7 @@
     number = {58},
     pages = {3024},
     author = {Adam Richie-Halford and Manjari Narayan and Noah Simon and Jason Yeatman and Ariel Rokem},
-    title = {Groupyr: Sparse Group Lasso in Python},
+    title = {Groupyr: Sparse Group Lasso in {P}ython},
     journal = {Journal of Open Source Software}
 }
 
@@ -221,7 +221,7 @@
 
 @article{Zhu:2022,
     author  = {Jin Zhu and Xueqin Wang and Liyuan Hu and Junhao Huang and Kangkang Jiang and Yanhang Zhang and Shiyun Lin and Junxian Zhu},
-    title   = {abess: A Fast Best-Subset Selection Library in Python and R},
+    title   = {abess: A Fast Best-Subset Selection Library in {P}ython and {R}},
     journal = {Journal of Machine Learning Research},
     year    = {2022},
     volume  = {23},
@@ -314,7 +314,7 @@
 }
 
 @article{Zhong:2022,
-  title = {An \$\{\textbackslash ensuremath\{\textbackslash ell\}\}\_\{0\}\{\textbackslash ensuremath\{\textbackslash ell\}\}\_\{2\}\$-Norm Regularized Regression Model for Construction of Robust Cluster Expansions in Multicomponent Systems},
+  title = {An ℓ 0 ℓ 2-Norm Regularized Regression Model for Construction of Robust Cluster Expansions in Multicomponent Systems},
   author = {Zhong, Peichen and Chen, Tina and Barroso-Luque, Luis and Xie, Fengyu and Ceder, Gerbrand},
   year = {2022},
   journaltitle = {Physical Review B},
