+++ paper/paper.bib
@@ -314,7 +314,7 @@
 }
 
 @article{Zhong:2022,
-  title = {An ℓ 0 ℓ 2-Norm Regularized Regression Model for Construction of Robust Cluster Expansions in Multicomponent Systems},
+  title = {An L0 L2-Norm Regularized Regression Model for Construction of Robust Cluster Expansions in Multicomponent Systems},
   author = {Zhong, Peichen and Chen, Tina and Barroso-Luque, Luis and Xie, Fengyu and Ceder, Gerbrand},
   year = {2022},
   journaltitle = {Physical Review B},
