### 2023-11-24 #5867 @jbytecode
@lbluque - I've just sent a pull request that includes some minor changes in both the bibtex and the manuscript. Could you please review the changes and then merge if it is okay to you.

Note: In equation 1, after the squre root, number of pipes at the start does not match the ones at the end. Could you please review the equation 1 and correct the notation?

Please ping me when you've done with it. Thank you in advance.
