+++ .pre-commit-config.yaml
@@ -71,6 +71,6 @@ repos:
   - id: rst-inline-touching-normal
 
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v1.7.0'  # Use the sha / tag you want to point at
+    rev: 'v1.7.1'  # Use the sha / tag you want to point at
     hooks:
     -   id: mypy
