### 2023-11-29 #5867 @lbluque
> @lbluque - I've just sent a pull request that includes some minor changes in both the bibtex and the manuscript. Could you please review the changes and then merge if it is okay to you.
>
> Note: In equation 1, after the squre root, number of pipes at the start does not match the ones at the end (|g||βg||2). Could you please review the equation 1 and correct the notation?
>
> Please ping me when you've done with it. Thank you in advance.
>
> β∗=argminβ;||Xβ−y||22+(1−α)λ∑g∈G|g||βg||2+αλ||β||1

Thanks a for the detailed checks @jbytecode! I fixed the above and pushed a few edits to the paper. Please have a look when you get a chance.
