### 2023-11-30 #5867 @jbytecode
@lbluque - the paper looks better now, thank you.

Could you please perform these tasks:

- Create a tagged release that includes all changes (the latest form) of the software. The version should be in form of `vx.y.z`, e.g., `v1.0.3`.
- Create a Zenodo archive. In the archive, the title and the author list of the original paper should match. Do not use any other custom titles.
- Report me the version tag, zenodo archive url, and the DOI given by Zenodo here.

Meanwhile you are fulfilling the requirements, I will be taking a look at the paper again.

Thank you in advance.
