### 2023-12-05 #5867 @lbluque
Hi @jbytecode - here is the release and zenodo archive information:

- [release v0.5.2](https://github.com/CederGroupHub/sparse-lm/releases/tag/v0.5.2)
- [10.5281/zenodo.10246640](https://doi.org/10.5281/zenodo.10246640)
