### 2023-12-05 #5867 @jbytecode
@lbluque - Thank you for adding ORCIDs. Almost done. The reference list looks fine. Please have a full read of the compiled pdf and perform a full grammar check. Correct any issue if exists, and then ping me again.
