+++ .github/workflows/release.yml
@@ -35,7 +35,7 @@ jobs:
     if: github.event.inputs.release-pypi == 'true'
 
     steps:
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
         with:
           name: artifact
           path: dist
