### 2023-12-20 79dc921
final paper edits based on FX comments
